refactor: move stringifyMessage to format.ts, change i18n
