+++ src/renderer/src/i18n/locales/ru-ru.json
@@ -144,7 +144,7 @@
       "message.new.branch.created": "Новая ветка создана",
       "message.new.context": "Новый контекст",
       "message.quote": "Цитата",
-      "message.raw": "Исходный текст",
+      "message.text": "Текст",
       "message.regenerate.model": "Переключить модель",
       "message.useful": "Полезно",
       "navigation": {
